+++ packages/react-querybuilder/src/async-debug.ts
@@ -1,3 +1,3 @@
-export * from './hooks/useAsyncOptionList';
+export * from './hooks/useAsyncOptionList/index.debug';
 export * from './redux/asyncOptionListsSlice';
 export type { AsyncOptionListsSliceState, CachedOptionList } from './redux/types';