+++ packages/react-querybuilder/src/hooks/useAsyncOptionList/useAsyncOptionList.test.tsx
@@ -6,17 +6,17 @@ import { waitABeat } from '@rqb-testing';
 import { act, renderHook } from '@testing-library/react';
 import * as React from 'react';
 import { Provider } from 'react-redux';
-import { QueryBuilderStateContext } from '../redux';
+import { QueryBuilderStateContext } from '../../redux';
 import {
   asyncOptionListsSlice,
   DEFAULT_CACHE_TTL,
   getOptionListsAsync,
-} from '../redux/asyncOptionListsSlice';
-import { queriesSlice } from '../redux/queriesSlice';
-import { warningsSlice } from '../redux/warningsSlice';
-import type { Schema, ValueEditorProps, VersatileSelectorProps } from '../types';
-import type { UseAsyncOptionListParams } from './useAsyncOptionList';
-import { useAsyncOptionList } from './useAsyncOptionList';
+} from '../../redux/asyncOptionListsSlice';
+import { queriesSlice } from '../../redux/queriesSlice';
+import { warningsSlice } from '../../redux/warningsSlice';
+import type { Schema, ValueEditorProps, VersatileSelectorProps } from '../../types';
+import { useAsyncOptionList } from './index';
+import type { UseAsyncOptionListParams } from './types';
 
 const defaultSchema = {
   suppressStandardClassnames: false,